requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.4.1=pyhd8ed1ab_0
ca-certificates=2021.5.30=h5b45459_0
django=3.1.8=pypi_0
django-cors-headers=3.7.0=pyh44b312d_0
djangorestframework=3.12.0=pypi_0
git=2.33.0=h57928b3_0
openssl=1.1.1k=h8ffe710_1
pip=21.2.4=pyhd8ed1ab_0
psycopg2=2.9.1=pypi_0
pyjwt=2.1.0=pyhd8ed1ab_0
python=3.9.6=h7840368_1_cpython
python_abi=3.9=2_cp39
pytz=2021.1=pyhd8ed1ab_0
setuptools=57.4.0=py39hcbf5309_0
sqlite=3.36.0=h8ffe710_0
sqlparse=0.4.1=pyh9f0ad1d_0
typing_extensions=3.10.0.0=pyha770c72_0
tzdata=2021a=he74cb21_1
ucrt=10.0.20348.0=h57928b3_0
vc=14.2=hb210afc_5
vs2015_runtime=14.29.30037=h902a5da_5
wheel=0.37.0=pyhd8ed1ab_1
